Change pool count type from i32 to i64 in self_stat.rs

+++ binaries/geph5-broker/src/self_stat.rs
@@ -27,7 +27,7 @@ pub async fn self_stat_loop() -> anyhow::Result<()> {
                 load_avg / available_parallelism().unwrap().get() as f64,
             )?;
         }
-        let pool_counts: Vec<(String, i32)> =
+        let pool_counts: Vec<(String, i64)> =
             sqlx::query_as("select pool,count(listen) from bridges_new group by pool")
                 .fetch_all(&*POSTGRES)
                 .await?;
